zigbee: zigbee: RI + CT update - Updated schema files and parsers to be able to use:    - Hex or base 10 numbers    - IEEE addresses are now hex string (size 16 hex digits). - Updated README.md file - Added logs into the test cases - Testing service properties of ZigBeeHost, ZigBeeNode and ZigBeeEndpoint services. - Cross check of the above service properties with the information available through the Service objects getters. - Changed command identifier data type from int to short. - Changed some input and output cluster properties types from Integer[]to int[].

+++ org.osgi.impl.service.zigbee/src/org/osgi/impl/service/zigbee/basedriver/ZCLFrameImpl.java
@@ -52,7 +52,7 @@ public class ZCLFrameImpl extends ZigBeeSerializer implements ZCLFrame {
 	 * 
 	 * @param commandId The command identifier of the ZCL Frame.
 	 */
-	public ZCLFrameImpl(int commandId) {
+	public ZCLFrameImpl(short commandId) {
 		super();
 		this.zclHeader = new ZCLHeaderImpl(commandId, false, true, false, (byte) 0);
 	}
