+++ org.osgi.impl.service.zigbee/src/org/osgi/impl/service/zigbee/basedriver/ZCLHeaderImpl.java
@@ -28,7 +28,7 @@ import org.osgi.service.zigbee.ZCLHeader;
  */
 public class ZCLHeaderImpl implements ZCLHeader {
 
-	private int		commandId;
+	private short	commandId;
 
 	private int		manufacturerCode			= -1;
 
@@ -60,7 +60,7 @@ public class ZCLHeaderImpl implements ZCLHeader {
 	 *
 	 */
 
-	public ZCLHeaderImpl(int commandId, boolean isClusterSpecificCommand, boolean isClientServerDirection,
+	public ZCLHeaderImpl(short commandId, boolean isClusterSpecificCommand, boolean isClientServerDirection,
 			boolean disableDefaultResponse, byte sequenceNumber) {
 		this.isClusterSpecificCommand = isClusterSpecificCommand;
 		this.isClientServerDirection = isClientServerDirection;
@@ -89,7 +89,7 @@ public class ZCLHeaderImpl implements ZCLHeader {
 	 * 
 	 * @param manufacturerCode The manufacturer code of the command.
 	 */
-	public ZCLHeaderImpl(int commandId, boolean isClusterSpecificCommand, boolean isClientServerDirection,
+	public ZCLHeaderImpl(short commandId, boolean isClusterSpecificCommand, boolean isClientServerDirection,
 			boolean disableDefaultResponse, byte sequenceNumber, int manufacturerCode) {
 
 		this(commandId, isClusterSpecificCommand, isClientServerDirection, disableDefaultResponse, sequenceNumber);
@@ -98,7 +98,7 @@ public class ZCLHeaderImpl implements ZCLHeader {
 		this.manufacturerCode = manufacturerCode;
 	}
 
-	public int getCommandId() {
+	public short getCommandId() {
 		return this.commandId;
 	}
